Tinathunight

+++ app/api/channel_routes.py
@@ -5,10 +5,8 @@ from flask_login import current_user, login_required
 
 channel_routes = Blueprint('channels', __name__)
 
-#add login_required
 
 @channel_routes.route('/<int:serversId>/all')
-# @login_required
 def all_channel(serversId):
   channels = Channel.query.filter_by(serverId=serversId).all()
   print(channels)
@@ -18,19 +16,18 @@ def all_channel(serversId):
   return channel_list
 
 @channel_routes.route('/new', methods=["GET","POST"])
-# @login_required
+@login_required
 def create_new_channel():
 
-    # creator_id = current_user.get_id()
-    # print(creator_id)v``
-    # print(server_id)
-    # curr_user_servers = Server.query.filter_by(creatorId=creator_id)
-
-    # if not curr_user_servers:
-    #     return {"error": "No server found for this user."}, 400
-
+    # auth REQUIRED, CURRENT USER  SERVER OWNER
     form = ChannelForm()
     form['csrf_token'].data = request.cookies['csrf_token']
+
+    server = Server.query.get(form.data["serverId"])
+    if server.creatorId != current_user.id:
+        return {'errors': {'message': 'Unauthorized'}}, 401
+
+
     if form.validate_on_submit():
         new_channel = {
             "name": form.data["name"],
@@ -45,8 +42,14 @@ def create_new_channel():
     return form.errors, 401
 
 @channel_routes.route('/<int:channelId>/delete', methods=['GET'])
-# @login_required
+@login_required
 def delete_channel(channelId):
+        # auth REQUIRED, CURRENT USER  SERVER OWNER
+    form = ChannelForm()
+    server = Server.query.get(form.data["serverId"])
+    if server.creatorId != current_user.id:
+        return {'errors': {'message': 'Unauthorized'}}, 401
+
     channel_to_delete=Channel.query.get(channelId)
     db.session.delete(channel_to_delete)
     db.session.commit()
@@ -54,10 +57,16 @@ def delete_channel(channelId):
 
 
 @channel_routes.route('/<int:channelId>/edit', methods=["POST"])
-# @login_required
+@login_required
 def update_channel(channelId):
+        # auth REQUIRED, CURRENT USER  SERVER OWNER
     form = ChannelForm()
     form['csrf_token'].data = request.cookies['csrf_token']
+
+    server = Server.query.get(form.data["serverId"])
+    if server.creatorId != current_user.id:
+        return {'errors': {'message': 'Unauthorized'}}, 401
+
     if form.validate_on_submit():
         channel_to_update=Channel.query.get(channelId)
         channel_to_update.name = form.data["name"]
